+++ src/stdatamodels/jwst/transforms/converters/jwst_models.py
@@ -20,19 +20,31 @@ class NIRCAMGrismDispersionConverter(TransformConverterBase):
     def from_yaml_tree_transform(self, node, tag, ctx):
         from stdatamodels.jwst.transforms import models
 
-        _fname = getattr(models, node["class_name"])
+        _fname = getattr(models, node["model_type"])
         return _fname(list(node['orders']),
                       list(node['lmodels']),
                       list(node['xmodels']),
                       list(node['ymodels']),
+                      list(node['inv_lmodels']),
+                      list(node['inv_xmodels']),
+                      list(node['inv_ymodels']),
                       )
 
     def to_yaml_tree_transform(self, model, tag, ctx):
+        xll = [list(m) for m in model.xmodels]
+        yll = [list(m) for m in model.ymodels]
+        lll = [list(m) for m in model.lmodels]
+        inv_xll = [list(m) for m in model.inv_xmodels]
+        inv_yll = [list(m) for m in model.inv_ymodels]
+        inv_lll = [list(m) for m in model.inv_lmodels]
         node = {'orders': list(model.orders),
-                'lmodels': list(model.lmodels),
-                'xmodels': list(model.xmodels),
-                'ymodels': list(model.ymodels),
-                'class_name': type(model).name
+                'xmodels': xll,
+                'ymodels': yll,
+                'lmodels': lll,
+                'inv_xmodels': inv_xll,
+                'inv_ymodels': inv_yll,
+                'inv_lmodels': inv_lll,
+                'model_type': type(model).name
                 }
         return node